fixed i guess

+++ src/chess/piece.rs
@@ -2,7 +2,7 @@ use core::panic;
 use std::convert::From;
 use std::fmt;
 
-#[derive(Debug, Eq, PartialEq)]
+#[derive(Debug, Eq, PartialEq,Copy,Clone)]
 pub struct Piece {
     pub p_type: PieceType,
     pub p_color: Color,
@@ -65,12 +65,12 @@ impl From<char> for PieceType {
 
 #[derive(Debug, Eq, PartialEq, Copy,Clone)]
 pub struct Position {
-    pub x: isize,
-    pub y: isize,
+    pub x: i8,
+    pub y: i8,
 }
 
 impl Position {
-    pub fn new(x: isize, y: isize) -> Self {
+    pub fn new(x: i8, y: i8) -> Self {
         Position { x, y }
     }
 }
